Added Package Controller and Some Basic Tweaks Done
Added short time hash to the image name of top header to make sure that every time user updates the photo browser loads the latest one.

Added a controller for packages and also created a migration for the packages.

Updated the email template which makes it more responsive.

Updated the routes for the admin pages.